CORE-1181 Migrate confidential identities project to build using the corda 5 API
Updated project to build using latest snapshot of corda5. This included upgrading to Java 11, Kotlin 1.4, Gradle 6.8.3, and the latest corda gradle plugin to allow cordapps to build as CPKs. I've also incremented the target release number from 1.0 to 2.0 since it's building against a major number increment of corda.

Before releasing project it will be necessary to:

* Set the version of corda5 to the release artifact when available
* Set the target platform version for cordapp configurations (https://r3-cev.atlassian.net/browse/CORE-1374)
* Set the version of the corda gradle plugin to the release artifact when available
* Remove the corda-os-maven-unstable repo from the build files

Remaining work will be to migrate the old DriverDSL/MockNetwork tests: https://r3-cev.atlassian.net/browse/CORE-1373
The existing tests have been removed in this PR to be replaced later, and a single e2e test was added just to verify the flows can be called successfully after the migration. Proper unit tests will be added in a follow on PR.